requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
chromadb==0.4.24
openai==1.12.0
python-dotenv==1.0.1
Werkzeug==3.0.1
